Move splash ignition to the main ignition
