+++ cspell.json
@@ -65,8 +65,7 @@
     "Vendicated",
     "Webauthn",
     "weblate",
-    "XLARGE",
-    "splashrenderer"
+    "XLARGE"
   ],
   "ignoreWords": [],
   "import": [],
