+++ src/renderer/index.ts
@@ -2,5 +2,10 @@ import * as replugged from "./replugged";
 
 window.replugged = replugged;
 
-await replugged.plugins.loadAll();
-await replugged.ignition.ignite();
+// Splash screen
+if (document.title === "Discord Updater") {
+  await replugged.ignition.startSplash();
+} else {
+  await replugged.plugins.loadAll();
+  await replugged.ignition.ignite();
+}
