--- src/splashrenderer/index.ts
@@ -1,3 +1,0 @@
-import * as ignition from "./managers/ignition";
-
-await ignition.ignite();
